requirements.txt
flask==2.2.2
flask-cors==3.0.10
Pillow==9.0.0
tensorflow==2.7.0